chall3/epicmail3/requirements.txt
atomicwrites==1.3.0
attrs==18.2.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
coverage==4.5.2
coveralls==1.6.0
docopt==0.6.2
entrypoints==0.3
flake8==3.7.7
Flask==1.0.2
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==6.0.0
pkg-resources==0.0.0
pluggy==0.9.0
psycopg2-binary==2.7.7
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
PyJWT==1.7.1
pytest==4.3.0
pytest-cov==2.6.1
python-dotenv==0.10.1
requests==2.21.0
six==1.12.0
urllib3==1.24.1
Werkzeug==0.14.1
